Basic Architecture Set up
Created the rough architecture of objects and inheritance of how I want my project to work. Borrowed a lot from the text adventure in terms of how commands are going to be issued.
Created Main class CreatureWar
Created abstract creature
Created Commands on how creatures can be arranged and battled.
Created a GameWorld for the Soldiers and Environment of the armies warring to be simulated in.