refix
Didn't realize main(String[] args) wasnt launching in CreatureWar class until I changed computers and cloned it. It was getting hung up on what was a verified creature so I refined it a bit and now I just need to allow them to engage and finish up the commands. then ill implement the tests.
I implemented an Enum class for Creature testing. I dont think this is the most efficient way but I want to show I understand enums a little and if I can I want to implement a way to use the enum to offer descriptions to user without calling a creature class.)